+++ package.json
@@ -22,6 +22,7 @@
     "@vitejs/plugin-react": "4.3.1",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "eslint": "^8.57.0",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.34.3",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
@@ -37,5 +38,11 @@
     "test": "vitest",
     "test:run": "vitest run"
   },
-  "proxy": "http://localhost:8080"
+  "proxy": "http://localhost:8080",
+  "overrides": {
+    "@babel/helpers": "^7.26.10",
+    "@babel/runtime": "^7.26.10",
+    "nanoid": "^3.3.8",
+    "micromatch": "^4.0.8"
+  }
 }
