+++ src/api/egovFetch.jsx
@@ -2,11 +2,29 @@ import { SERVER_URL } from "../config";
 
 import URL from "@/constants/url";
 import CODE from "@/constants/code";
+import { getSessionItem, setSessionItem } from "@/utils/storage";
+
+export function getQueryString(params) {
+  return `?${Object.entries(params)
+    .map((e) => e.join("="))
+    .join("&")}`;
+}
 
 export function requestFetch(url, requestOptions, handler, errorHandler) {
   console.groupCollapsed("requestFetch");
   console.log("requestFetch [URL] : ", SERVER_URL + url);
   console.log("requestFetch [requestOption] : ", requestOptions);
+
+  // Login 했을경우 JWT 설정
+  const sessionUser = getSessionItem("loginUser");
+  const sessionUserId = sessionUser?.id || null;
+  const jToken = getSessionItem("jToken");
+  if (sessionUserId != null && sessionUserId !== undefined) {
+    if (!requestOptions["headers"]) requestOptions["headers"] = {};
+    if (!requestOptions["headers"]["Authorization"])
+      requestOptions["headers"]["Authorization"] = null;
+    requestOptions["headers"]["Authorization"] = jToken;
+  }
 
   //CORS ISSUE 로 인한 조치 - origin 및 credentials 추가
   // origin 추가
@@ -27,7 +45,7 @@ export function requestFetch(url, requestOptions, handler, errorHandler) {
     .then((resp) => {
       if (Number(resp.resultCode) === Number(CODE.RCV_ERROR_AUTH)) {
         alert("Login Alert"); //index.jsx라우터파일에 jwtAuthentication 함수로 공통 인증을 사용하는 코드 추가로 alert 원상복구
-        sessionStorage.setItem("loginUser", JSON.stringify({ id: "" }));
+        setSessionItem("loginUser", { id: "" });
         window.location.href = URL.LOGIN;
         return false;
       } else {
