+++ src/constants/code.jsx
@@ -18,5 +18,4 @@ const CODE = {
     DATE_DAY            : "day",
 }
 
-// eslint-disable-next-line react-refresh/only-export-components
 export default CODE;
