+++ src/pages/admin/manager/EgovAdminPasswordUpdate.jsx
@@ -74,7 +74,6 @@ function EgovAdminPasswordUpdate(props) {
   };
 
   useEffect(() => {
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   console.log("------------------------------EgovAdminPasswordUpdate [End]");
