+++ vite.config.js
@@ -4,7 +4,7 @@ import react from "@vitejs/plugin-react-swc";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react({
-    // JSX 처리를 위한 추가 설정
+    // Include JSX processing for JS and JSX files
     include: "**/*.{jsx,js}",
   })],
   
@@ -21,16 +21,16 @@ export default defineConfig({
     environment: "jsdom",
     setupFiles: "./vitest.setup.js",
     transformMode: {
-      web: [/\.[jt]sx?$/],  // 모든 JS/JSX/TS/TSX 파일을 web 모드로 변환
+      web: [/\.[jt]sx?$/],  // Transform all JS/JSX/TS/TSX files in web mode
     },
   },
   build: {
     chunkSizeWarningLimit: 100000000,
   },
-  // .js 파일에서 JSX 구문을 지원하도록 설정
+  // Enable JSX syntax in .js files
   esbuild: {
     loader: "jsx",
-    include: /\.[jt]sx?$/,  // .js, .jsx, .ts, .tsx 모두 포함
+    include: /\.[jt]sx?$/,  // Include all .js, .jsx, .ts, .tsx files
     exclude: [],
   },
   optimizeDeps: {
